+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 ## Unreleased
 
+### Added
+
+- Added animations to `<NormalizedStackedBarChart />`.
+
 ## [0.23.0] - 2021-10-27
 
 ### Added
 
 - Added `<HorizontalBarChart />`.
 - Added `grid.showHorizontalLines` theme option to `<StackedAreaChart />`.
-### Fixed
-
-- Better print support for line chart, bar charts and area chart. Charts should now be responsive to print on most browsers. On Firefox, the chart from the browser window will not overflow the print page.
 
 ## [0.22.0] - 2021-10-22